Basic viewer integration
