+++ appinfo/app.php
@@ -37,14 +37,13 @@ $eventDispatcher = \OC::$server->getEventDispatcher();
 $eventDispatcher->addListener(
 	'OCA\Files::loadAdditionalScripts',
 	function() {
-		\OCP\Util::addScript('richdocuments', 'viewer');
-		\OCP\Util::addStyle('richdocuments', 'viewer');
+		\OCP\Util::addScript('richdocuments', 'files');
 	}
 );
 $eventDispatcher->addListener(
 	'OCA\Files_Sharing::loadAdditionalScripts',
 	function() {
-		\OCP\Util::addScript('richdocuments', 'viewer');
+		\OCP\Util::addScript('richdocuments', 'files');
 	}
 );
